+++ httpserver/src/main/kotlin/pt/isel/ps/g06/httpserver/dataAccess/db/dto/DbNutritionalDto.kt
@@ -1,7 +1,7 @@
 package pt.isel.ps.g06.httpserver.dataAccess.db.dto
 
 open class DbNutritionalDto(
-        val carbs: Int,
-        val amount: Int,
+        val carbs: Float,
+        val amount: Float,
         val unit: String
 )
