+++ httpserver/src/main/kotlin/pt/isel/ps/g06/httpserver/dataAccess/db/repo/InsulinProfileDbRepository.kt
@@ -52,9 +52,9 @@ class InsulinProfileDbRepository(
             profileName: String,
             startTime: LocalTime,
             endTime: LocalTime,
-            glucoseObjective: Int,
-            insulinSensitivityFactor: Int,
-            carbohydrateRatio: Int
+            glucoseObjective: Float,
+            insulinSensitivityFactor: Float,
+            carbohydrateRatio: Float
     ): DbUserInsulinProfileDto {
         return databaseContext.inTransaction { handle ->
             val insulinProfileDao = handle.attach(insulinProfileDaoClass)
